+++ template-sample-app-3/template/src/utils/formatter.js
@@ -14,6 +14,9 @@ import '@formatjs/intl-numberformat/polyfill';
 import '@formatjs/intl-numberformat/locale-data/en';
 
 const DEFAULT_LOCALE = 'en-US';
+const CIRCLE_SVG = `<svg width="10px" height="10px" fill="#47db99" viewBox="0 0 10 10" xmlns="http://www.w3.org/2000/svg">
+  <circle cx="5" cy="5" r="5"/>
+</svg>`;
 
 export const currencyFormatter = (value) => {
   const _value = parseLogicalOperation(value);
@@ -41,6 +44,13 @@ export const numberFormatter = (value) => {
   );
 };
 
+export const intervalsFormatter = (value, dataIndex, ticks) => {
+  const _value = numberFormatter(value);
+  if (!ticks || dataIndex === undefined) return _value;
+  const intervals = moneyInterval(dataIndex, ticks);
+  return `${intervals} <br/> ${CIRCLE_SVG} ${_value}`;
+};
+
 const parseLogicalOperation = (value) => {
   if (!isNaN(value)) return { value, operation: '' };
 
@@ -59,5 +69,27 @@ const parseLogicalOperation = (value) => {
     return isNaN(value) ? { value: 0, operation: '' } : { value, operation };
   } catch {
     throw new Error(`You are using a numberFormatter on a not valid value: ${value}`);
+  }
+};
+
+const moneyInterval = (dataIndex, ticks) => {
+  const isFirst = dataIndex === 0;
+  try {
+    if (isFirst || dataIndex === ticks.length) {
+      const comparison = isFirst ? '<' : '≥';
+      const formattedValue = isFirst
+        ? currencyFormatter(ticks[dataIndex])
+        : currencyFormatter(ticks[dataIndex - 1]);
+      return `${comparison} ${formattedValue.prefix}${formattedValue.value}`;
+    } else {
+      dataIndex = dataIndex - 1;
+      const prevTick = currencyFormatter(ticks[dataIndex]);
+      const nextTick = currencyFormatter(ticks[dataIndex + 1]);
+      return `${prevTick.prefix}${prevTick.value} — ${nextTick.prefix}${nextTick.value}`;
+    }
+  } catch {
+    throw new Error(
+      `You are using an "intervalsFormatter" on a not valid index: ${dataIndex} & for the ticks ${ticks}`
+    );
   }
 };