+++ hygen/_templates/layer/new/inject_view_import_actionsSlice.ejs.t
@@ -2,6 +2,11 @@
 inject: true
 to: "<%= attach ? `src/${view_path}` : null %>"
 after: "from 'react'"
-skip_if: \{ addLayer,
+skip_if: addLayer,
 ---
-import { addLayer, removeLayer, addSource, removeSource } from '@carto/react-redux';
+import {
+  addLayer,
+  removeLayer,
+  addSource,
+  removeSource,
+} from '@carto/react-redux';