+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Not released
 
+- Bump carto-react 1.2.1-alpha.1[#316](https://github.com/CartoDB/carto-react-template/pull/317)
 - Add DrawingTool in template-sample-app-3 [#314](https://github.com/CartoDB/carto-react-template/pull/314)
 - Preparing alpha 1.2.0 version of the template [#313](https://github.com/CartoDB/carto-react-template/pull/313)
 - Bump 8.7.0-alpha.11 [#316](https://github.com/CartoDB/carto-react-template/pull/316)