+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Not released
 
+## 1.2.0
+
+### (prerelease) 1.2.0-beta.0 (2022-02-02)
+
 - Bump versions: carto-react 1.2.1-beta.1 and deck.gl 8.7.0-beta.2 [#320](https://github.com/CartoDB/carto-react-template/pull/320)
 - Update template-base-3-typescript template to deck.gl 8.7.0-alpha.11 [#318](https://github.com/CartoDB/carto-react-template/pull/318)
 - Bump carto-react 1.2.1-alpha.1 [#317](https://github.com/CartoDB/carto-react-template/pull/317)